python_libraries/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==4.8.0
    # via httpx
black==22.12.0
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via black
coverage[toml]==7.7.0
    # via pytest-cov
docopt==0.6.2
    # via -r requirements.in
duckdb==1.2.1
    # via -r requirements.in
faker==37.0.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==2.10
    # via
    #   anyio
    #   httpx
iniconfig==1.1.1
    # via pytest
jinja2==2.11.3
    # via -r requirements.in
markupsafe==1.1.1
    # via jinja2
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
psutil==7.0.0
    # via -r requirements.in
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.25.3
    # via -r requirements.in
pytest-cov==6.0.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
redis==5.2.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
typing-extensions==4.13.2
    # via anyio
tzdata==2025.1
    # via faker
